Add new option to download online stream till end. It will download current stream, then every 5 minutes update it and download new part. Also able to auto split it (when enought video downloaded, will convert it and continue waiting)

+++ TwitchLeecher/TwitchLeecher.Services/Services/ProcessingService.cs
@@ -37,7 +37,7 @@ namespace TwitchLeecher.Services.Services
 
         #region Methods
 
-        public void ConcatParts(Action<string> log, Action<string> setStatus, Action<double> setProgress, VodPlaylist vodPlaylist, string concatFile)
+        public void ConcatParts(Action<string> log, Action<string> setStatus, Action<double> setProgress, VodPlaylist vodPlaylist, string concatFile, bool removeFiles = true)
         {
             setStatus("Merging files");
             setProgress(0);
@@ -63,7 +63,8 @@ namespace TwitchLeecher.Services.Services
                         }
                     }
 
-                    FileSystem.DeleteFile(part.LocalFile);
+                    if (removeFiles)
+                        FileSystem.DeleteFile(part.LocalFile);
 
                     setProgress(i * 100 / partsCount);
                 }
